update long parameters to big integer

+++ src/main/java/com/brkygngr/bracketcoco/dto/BracketCombinationResponseDto.java
@@ -4,19 +4,20 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import javax.validation.Valid;
 import javax.validation.constraints.PositiveOrZero;
+import java.math.BigInteger;
 
 @Valid
 public class BracketCombinationResponseDto {
 
   @PositiveOrZero
-  private final long totalCombinations;
+  private final BigInteger totalCombinations;
 
-  public BracketCombinationResponseDto(final long totalCombinations) {
+  public BracketCombinationResponseDto(final BigInteger totalCombinations) {
     this.totalCombinations = totalCombinations;
   }
 
   @JsonProperty("tc")
-  public long getTotalCombinations() {
+  public BigInteger getTotalCombinations() {
     return totalCombinations;
   }
 }
